new file:   content/declare-float/rust/b.rs

+++ content/declare-float/rust/b.rs
@@ -1,8 +1,8 @@
 fn main() {
    // example 1
-   let n1: f32 = 1e6;
+   let n1: f64 = 1e6;
    // example 2
-   let n2: f32 = 9.9;
+   let n2: f64 = 9.9;
    // print
    println!("{} {}", n1, n2);
 }
